baofd_api/requirements.txt
Flask == 3.0.3
scikit-learn == 1.5.2
pandas == 2.2.3
joblib == 1.4.2
boto3 == 1.35.49
python-dotenv == 1.0.1
pytest == 8.3.3
moto == 5.0.18
openpyxl == 3.1.5 
numpy == 2.1.2